收藏功能、迟书等价删除（未加入扣除用户时间）

+++ src/test/java/com/hyht/LateLetter/extraUtil/JobsTest.java
@@ -17,7 +17,7 @@ public class JobsTest {
         String dirStr = EnvirArgs.extraFilePath + "\\checkImg";
         File file = new File(dirStr);
         File[] files = file.listFiles();
-        if(files.length > 0){
+        if(files != null && files.length > 0){
             for(File temp : files){
                 if(temp.isFile()){
                     temp.delete();
@@ -29,4 +29,30 @@ public class JobsTest {
             System.out.println("nothing was found ");
         }
     }
+
+    @Test
+    public void deleteDir(){
+        String dirStr = EnvirArgs.extraFilePath + "\\letterExtraFile\\10005";
+        File file = new File(dirStr);
+        if(!file.exists()  && !file.isDirectory()){
+            File[] files = file.listFiles();
+            if(files != null && files.length > 0){
+                for(File temp : files){
+                    if(temp.isFile()){
+                        temp.delete();
+                    }
+                }
+                System.out.println("bitch i wil fuck you ");
+                //删除目录
+                file.delete();
+            }
+            else{
+                System.out.println("nothing was found ");
+            }
+        }
+        else{
+            System.out.println("file was found ");
+        }
+    }
+
 }
